% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipopCpp_main.R
\docType{package}
\name{ipopCpp-package}
\title{Rcpp Quadratic Problem Solver}
\description{
Rcpp implementation of the kernlab::ipop() quadratic
problem solver
}
\details{
The only function you're likely to need from \pkg{ipopCpp} is
\code{\link{ipopCpp}}.
}

