% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pmax_arma}
\alias{pmax_arma}
\title{RcppArmadillo implemenation of the r pmax function}
\usage{
pmax_arma(x, bound)
}
\arguments{
\item{x}{the (aramdillo) vector}

\item{bound}{the (double) bound}
}
\description{
RcppArmadillo implementation of the R pmax()
function
}

