% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solution_v_forc_example}
\alias{solution_v_forc_example}
\title{Synthetic Control exmaple of a solution for the variable weights}
\format{A vector of 12 elements}
\usage{
solution_v_forc_example
}
\description{
Synthetic Control exmaple of a solution for the variable weights
}
\keyword{datasets}

