% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X_scaled_forc_example}
\alias{X_scaled_forc_example}
\title{Synthetic control X matrix of scaled predictor data}
\format{A 12 x 41 matrix}
\usage{
X_scaled_forc_example
}
\description{
Synthetic control X matrix of scaled predictor data
}
\keyword{datasets}

