% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Z_forc_example}
\alias{Z_forc_example}
\title{Synthetic control Z matrix of outcome variables}
\format{A 18 x 42 matrix}
\usage{
Z_forc_example
}
\description{
Synthetic control Z matrix of outcome variables
}
\keyword{datasets}

